feat: parse monthlyUsage.dailyServiceUsages[].date as Date
This field does not end with "At", which means it is ignored by the
existing parseDateFields helper. This PR adjusts this helper to
support parsing other fields as well.

To avoid accidental braking changes by applying the new parser on all
existing data structures, the helper now has a new optional
shouldParseField() param to identify other fields to be parsed. Without
this param, the helper behaves just like before.