feat: add forcePermissionLevel run option (#743)
We want Actor developers to be able to easily test their full permission
Actors with limited permissions without changing the Actor configuration
for everybody or redeploying the Actor under a different name. For that
reason, we have introduced a new `forcePermissionLevel` run option. This
PR adds the option to the client.

Full context here: https://github.com/apify/apify-core/pull/22681