feat: allowed signature to be passed in kv-store/datasets (#761)
When working with output schema in UI I found that we do not allow
passing signature in client to read data from storages. This PR does not
fully resolve it, but it resolves most of the issues I found.

I will leave the rest to Daniil, since he might now better which
endpoints use signatures and which do not.