feat: add startedBefore and startedAfter to run list (#763)
This PR adds new options to RunCollectionClient.list(): `startedBefore`
and `startedAfter`.

Note: this is the first time we use dates in query in the API (correct
me if I am wrong), so in case we'll add any dates to query params in
future they should use the same logic. Based on agreement with @barjin
we decided to allow users to pass dates in 2 formats: as Date object or
as a string.